requirements.txt
confluent-kafka[avro]>=2.4.0
pymongo>=4.6.0
python-dotenv>=1.0.1
